Add double-click-text to add relationship and other text-mouse events
We can make relations between entities by clicking on text entities the same way as when interacting with nodes.
Mousing over and out of a text entity highlights the corresponding node.
Bugfix: Fixed an issue where cutting relations would cut the right relation in displayedGraph.relations, but the wrong one in editedGraph.relations