Node charge force depends on number of nodes displayed. Set initial node positions to make animations smoother.
Various adjustments. Changed node from global to local variable. Need to do the rest sometime.
Issue: When a same-as group has many nodes all stuck together, the group flies around beacuse of the collision force trying to keep the nodes from overlapping, as well as the charge force I think. We need to change node size dynamically, or change forces, or something. It's amusing to have same-as groups flying around.