Add an overlay div with adjustable size. The nodes shown in the graph correspond to the the text elements underneath the overlay div.
Right now, shrinking the overlay succesfully removes nodes and links. But enlarging it fails to create new nodes and links.
In order to do it we need some refactoring.
1. Change how entity subsets are determined. Right now we filter entities by predetermined filters. This is suboptimal because machine output occasionally creates "bad" links between entities - like making a person and a place spouses.
What we should do is the filtering, but also show any entities that are linked by the chosen relationship. The entities that would normally fail the filter should be marked as something for a person to check.
2. The nbx should highlight the same entities that correspond to displayed nodes. This will make things with the view box overlay easier. When the view box is moved around or resized, we want to hide/show only nodes that are linked by the chosen relationship.
3. Once these are in place, we can correctly rebuild the displayedGraph.