Change links from being global to local variable.
Issue: There is some unexpected behavior when relationship type "same as" is selected. The
links don't show up initially. If you select "same as" again, they show,
and if you do it again, they disappear again. The nodes behave as if
links exist, and the convex hulls are fine. So displayedGraph.relations
is fine. It's just that we're missing the "line" HTML elements.