Highlight node of bad rel works, delay sim start
The link for a bad relation only gets highlighted if the relationship
button is pressed twice. The bad node representing an invalid entity for
the chosen relationship type is successfully circled in red.
Added the delay timer to startSimulation in loadDocument() to make the
inital animation smoother. Frame rates drop as the file explorer window
closes and the simulation starts.